Refactor conftest.py
